git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/branches/hpx/release@7292 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ external/coroutine/boost/coroutine/detail/context_base.hpp
@@ -41,7 +41,7 @@
 #endif
 
 #ifdef BOOST_COROUTINE_USE_ATOMIC_COUNT
-#  include <boost/detail/atomic_count.hpp>
+#  include <boost/atomic.hpp>
 #endif
 #include <cstddef>
 #include <algorithm> //for swap
@@ -55,19 +55,48 @@
 #endif
 #include <boost/coroutine/exception.hpp>
 #include <boost/coroutine/detail/noreturn.hpp>
-namespace boost { namespace coroutines { namespace detail {
+#include <boost/assert.hpp>
+
+namespace boost { namespace coroutines { namespace detail
+{
+#ifdef BOOST_COROUTINE_USE_ATOMIC_COUNT
+  /////////////////////////////////////////////////////////////////////////////
+  //
+  struct allocation_counters
+  {
+      allocation_counters()
+      {
+          for(std::size_t i = 0; i < BOOST_COROUTINE_NUM_HEAPS; ++i)
+              m_allocation_counter[i].store(0);
+      }
+
+      boost::atomic_uint64_t& get(std::size_t i)
+      {
+          BOOST_ASSERT(i < BOOST_COROUTINE_NUM_HEAPS);
+          return m_allocation_counter[i];
+      }
+
+      boost::atomic_uint64_t m_allocation_counter[BOOST_COROUTINE_NUM_HEAPS];
+  };
+#endif
 
   /////////////////////////////////////////////////////////////////////////////
   const std::ptrdiff_t default_stack_size = -1;
 
   template<typename ContextImpl>
-  class context_base : public ContextImpl {
+  class context_base : public ContextImpl
+  {
   public:
+    enum deleter_mode
+    {
+        deleter_delete = 1,
+        deleter_reset = 2
+    };
 
     typedef ContextImpl context_impl;
     typedef context_base<context_impl> type;
     typedef boost::intrusive_ptr<type> pointer;
-    typedef void deleter_type(const type*);
+    typedef void deleter_type(type const*, deleter_mode);
 
     template<typename Derived>
         context_base(Derived& derived,
@@ -98,7 +127,7 @@ namespace boost { namespace coroutines { namespace detail {
       return count() == 1;
     }
 
-    std::size_t count() const {
+    boost::int64_t count() const {
       return m_counter;
     }
 
@@ -109,8 +138,12 @@ namespace boost { namespace coroutines { namespace detail {
     void release() const {
       BOOST_ASSERT(m_counter);
       if(--m_counter == 0) {
-        m_deleter(this);
+        m_deleter(this, deleter_delete);
       }
+    }
+
+    void reset() const {
+      m_deleter(this, deleter_reset);
     }
 
     void count_down() throw() {
@@ -347,13 +380,33 @@ namespace boost { namespace coroutines { namespace detail {
       } catch(...) {}
     }
 
-    static boost::uint64_t get_allocation_count()
+    static boost::uint64_t get_allocation_count_all()
     {
-      return m_allocation_counter;
+        boost::uint64_t count = 0;
+#ifndef BOOST_COROUTINE_USE_ATOMIC_COUNT
+        for (std::size_t i = 0; i < BOOST_COROUTINE_NUM_HEAPS; ++i)
+            count += m_allocation_counter[i];
+#else
+        for (std::size_t i = 0; i < BOOST_COROUTINE_NUM_HEAPS; ++i)
+            count += m_allocation_counters.get(i).load();
+#endif
+        return count;
     }
-    static boost::uint64_t increment_allocation_count()
+    static boost::uint64_t get_allocation_count(std::size_t heap_num)
     {
-      return ++m_allocation_counter;
+#ifndef BOOST_COROUTINE_USE_ATOMIC_COUNT
+        return m_allocation_counter[heap_num];
+#else
+        return m_allocation_counters.get(heap_num).load();
+#endif
+    }
+    static boost::uint64_t increment_allocation_count(std::size_t heap_num)
+    {
+#ifndef BOOST_COROUTINE_USE_ATOMIC_COUNT
+        return ++m_allocation_counter[heap_num];
+#else
+        return ++m_allocation_counters.get(heap_num);
+#endif
     }
 
   protected:
@@ -428,19 +481,25 @@ namespace boost { namespace coroutines { namespace detail {
     }
 
     template <typename ActualCtx>
-    static void deleter (const type* ctx)
+    static void deleter (type const* ctx, deleter_mode mode)
     {
-        ActualCtx::destroy(static_cast<ActualCtx*>(const_cast<type*>(ctx)));
+        if (deleter_delete == mode)
+            ActualCtx::destroy(static_cast<ActualCtx*>(const_cast<type*>(ctx)));
+        else if (deleter_reset == mode)
+            ActualCtx::reset(static_cast<ActualCtx*>(const_cast<type*>(ctx)));
+        else {
+            BOOST_ASSERT(deleter_delete == mode || deleter_reset == mode);
+        }
     }
 
     typedef typename context_impl::context_impl_base ctx_type;
     ctx_type m_caller;
 #ifndef BOOST_COROUTINE_USE_ATOMIC_COUNT
     mutable std::size_t m_counter;
-    static std::size_t m_allocation_counter;
+    static std::size_t m_allocation_counter[BOOST_COROUTINE_NUM_HEAPS];
 #else
-    mutable boost::detail::atomic_count m_counter;
-    static boost::detail::atomic_count m_allocation_counter;
+    mutable boost::atomic_uint64_t m_counter;
+    static allocation_counters m_allocation_counters;
 #endif
     deleter_type * m_deleter;
     context_state m_state;
@@ -457,8 +516,9 @@ namespace boost { namespace coroutines { namespace detail {
 #ifdef BOOST_COROUTINE_USE_ATOMIC_COUNT
   // initialize static allocation counter
   template <typename ContextImpl>
-  boost::detail::atomic_count context_base<ContextImpl>::m_allocation_counter(0);
+  allocation_counters context_base<ContextImpl>::m_allocation_counters;
 #endif
 
-} } }
+}}}
+
 #endif
